requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-multipart==0.0.20
opencv-python-headless==4.11.0.86
numpy==2.2.3
scikit-learn==1.6.1
scipy==1.15.2
Pillow==11.1.0
pydantic==2.10.6
pydantic-settings==2.8.0
python-dotenv==1.0.1
slowapi==0.1.9
a2wsgi==1.11.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.3
